- file added
This assignment revolved around a board game styled off the N-Queens program. The object of the game was to be the last player to put a piece (a queen) on the board, and therefore minimize the amount of queens the other player can play. The project uses a variety of Minimax algorithms, some with Alpha-Beta Pruning, some with Depth-Cutoff. There is also a large number of tests for the code present. Overall, it demonstrates a fine understanding of minimax and testing practices.

I don't think the project works perfectly, though it was graded well, but the general concepts of Minimax Search, Alpha-Beta Pruning, Depth-Cuttoff with Minimax, et. al. are there. As well, organization is lacking, but that was asked of us.